fix: return error when rel path is nil in source state instead of allowing later panic
The (*Config.targetState).targetRelPaths method has an option,
`mustBeInSourceState`. It looks like its intended to be used to return
an error when the target path has been removed from the source state.
But if the target path was never in the source state, the method will
not error because of it.

In many commands, subsequent to using this method with that option,
the target path is retrieved from the source state with
`(*SourceState).MustEntry` but this will panic if the target path
would have returned nil when using `(*SourceState).Get`.

This panic is user facing, running a command like
```
chezmoi merge /path/thats/not/in/source/state
```
will exhibit it.

Now, instead, the user will see something like
```
chezmoi: /path/thats/not/in/source/state: not in source state
```